+++ CHANGELOG.md
@@ -1,8 +1,15 @@
 # Changelog
 All notable changes to this package will be documented in this file.
 
-The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) as of release 4.0.0.
+The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
+
+## [3.0.0] - 2023-06-05
+This version is compatible with Unity 2022.3.0f1
+
+### Changed
+- ASCII now uses RTHandles and Blitter APIs
+
 
 ## [2.0.0] - 2023-05-18
 ### Added